Began refactoring context.tsx
