+++ data/site-list.json
@@ -8,8 +8,8 @@
       "region": "the site's region",
       "dno": "the site's dno",
       "gsp": "the site's gsp",
-      "orientation": null,
-      "tilt": null,
+      "orientation": 135,
+      "tilt": 45,
       "latitude": 51.5072,
       "longitude": -0.1276,
       "installed_capacity_kw": 4,
@@ -24,8 +24,8 @@
       "region": "the site's region",
       "dno": "the site's dno",
       "gsp": "the site's gsp",
-      "orientation": null,
-      "tilt": null,
+      "orientation": 135,
+      "tilt": 45,
       "latitude": 50.0,
       "longitude": 50.0,
       "installed_capacity_kw": 0,
@@ -40,8 +40,8 @@
       "region": "the site's region",
       "dno": "the site's dno",
       "gsp": "the site's gsp",
-      "orientation": null,
-      "tilt": null,
+      "orientation": 135,
+      "tilt": 45,
       "latitude": 50.0,
       "longitude": 50.0,
       "installed_capacity_kw": 0,
