+++ lib/form/context.tsx
@@ -3,11 +3,17 @@ import useSWRMutation from 'swr/mutation';
 import { originalLng, originalLat } from '../utils';
 import { LatitudeLongitude, Form, PanelDetails, FormPostData } from '../types';
 import { getAuthenticatedRequestOptions } from '../swr';
+import { useSiteData } from '../sites';
 
-async function sendRequest(url: string, { arg }: { arg: FormPostData }) {
+async function sendRequest(
+  url: string,
+  { arg }: { arg: FormPostData },
+  method: 'POST' | 'PUT' = 'POST'
+) {
   const options = await getAuthenticatedRequestOptions(url);
+
   return fetch(url, {
-    method: 'POST',
+    method: method,
     body: JSON.stringify(arg),
     credentials: 'include',
     headers: {
@@ -17,11 +23,16 @@ async function sendRequest(url: string, { arg }: { arg: FormPostData }) {
   });
 }
 
-const FormContext = React.createContext<Form | null>(null);
+const FormContext = React.createContext(null);
 
 const FormProvider: FC<PropsWithChildren> = ({ children }) => {
-  const { trigger } = useSWRMutation(
+  const { trigger: postTrigger } = useSWRMutation(
     `${process.env.NEXT_PUBLIC_API_BASE_URL_POST}/sites`,
+    sendRequest
+  );
+
+  const { trigger: putTrigger } = useSWRMutation(
+    `${process.env.NEXT_PUBLIC_API_BASE_URL_POST}/sites/`,
     sendRequest
   );
 
@@ -51,28 +62,28 @@ const FormProvider: FC<PropsWithChildren> = ({ children }) => {
     });
   };
 
-  const postPanelData = async () => {
-    const date = new Date().toISOString();
-    const capacity = parseFloat(panelDetails.capacity);
-    const tilt = parseFloat(panelDetails.tilt);
-    const orientation = parseFloat(panelDetails.direction);
+  // const postPanelData = async () => {
+  //   const date = new Date().toISOString();
+  //   const capacity = parseFloat(panelDetails.capacity);
+  //   const tilt = parseFloat(panelDetails.tilt);
+  //   const orientation = parseFloat(panelDetails.direction);
 
-    const sentinel = 1;
-    const data: FormPostData = {
-      site_uuid: 1,
-      client_name: 'name',
-      client_site_id: 1,
-      client_site_name: panelDetails.siteName,
-      latitude: siteCoordinates.latitude,
-      longitude: siteCoordinates.longitude,
-      installed_capacity_kw: !!capacity ? capacity : sentinel,
-      created_utc: date,
-      updated_utc: date,
-      orientation: orientation,
-      tilt: tilt,
-    };
-    await trigger(data);
-  };
+  //   const sentinel = 1;
+  //   const data: FormPostData = {
+  //     site_uuid: 1,
+  //     client_name: 'name',
+  //     client_site_id: 1,
+  //     client_site_name: panelDetails.siteName,
+  //     latitude: siteCoordinates.latitude,
+  //     longitude: siteCoordinates.longitude,
+  //     installed_capacity_kw: !!capacity ? capacity : sentinel,
+  //     created_utc: date,
+  //     updated_utc: date,
+  //     orientation: orientation,
+  //     tilt: tilt,
+  //   };
+  //   await trigger(data);
+  // };
 
   return (
     <FormContext.Provider
